Fase0/tests/requirements.txt
pytest==8.1.1
pytest-asyncio==0.23.6
asyncpg==0.29.0
redis==5.0.4
minio==7.2.7
qdrant-client==1.9.1
httpx==0.27.0
python-dotenv==1.0.1
